[stack-manager] Optimize how we obtain the runtime jars for external JVM processes
